pyproject.toml
[project]
name = "workhuman-ml-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "boto3>=1.42.19",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "watchtower>=3.4.0",
    "redis>=5.2.1",
]
